UBUNTU: SAUCE: tools -- add ability to disable libbfd
We do not want to be linked to libbfd as this is a tightly versioned
package which does not maintain its ABI.  This prevents us from have
multiple tools packages installed.  Turn that off and we will fallback
to libiberty.

BugLink: http://bugs.launchpad.net/bugs/1748922
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>

[fix add ability to disable libbfd]
BugLink: https://bugs.launchpad.net/bugs/1826410
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
[ arighi: context adjustments for 6.4-rc1 ]
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>
[ tjaalton: adjustments for 6.12-rc1]
Signed-off-by: Timo Aaltonen <timo.aaltonen@canonical.com>