ALSA: hda/realtek: Autodetect Cirrus Logic companion amplifier bindings
BugLink: https://bugs.launchpad.net/bugs/2089889

We do not need model specific HDA quirks to construct the component
binding details of Cirrus Logic companion amplifiers as this information
is already present in the ACPI.

Quirks are then only required for special workarounds not described in
the ACPI such as internal configuration of the Realtek codecs.

The codec pointer is now initialized in hda_component_manager_init() so
that we can detect when companion amplifiers are found in the ACPI but
the SSID invokes a quirk that also attempts to create the component
binding.

Signed-off-by: Simon Trimmer <simont@opensource.cirrus.com>
Link: https://patch.msgid.link/20240829161114.140938-1-simont@opensource.cirrus.com
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(backported from commit 2186fe21e57aada45f365a79cdb6d21140a539e2)
[Chris Chiu: drop quirks which are not landed on 6.11 yet]
Signed-off-by: Chris Chiu <chris.chiu@canonical.com>
Acked-by: Chia-Lin Kao (AceLan) <chia-lin.kao@canonical.com>
Signed-off-by: Kuan-Ying Lee <kuan-ying.lee@canonical.com>