ASoC: sdw_utils: Add support for exclusion DAI quirks
BugLink: https://bugs.launchpad.net/bugs/2086250

The system contains a mechanism for certain DAI links to be included
based on a quirk. Add support for certain DAI links to excluded based on
a quirk, this is useful in situations where the vast majority of SKUs
utilise a feature so it is easier to quirk on those that don't.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Reviewed-by: Liam Girdwood <liam.r.girdwood@intel.com>
Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Link: https://patch.msgid.link/20241016030344.13535-2-yung-chuan.liao@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 3d9b44d0972be1298400e449cfbcc436df2e988e)
Signed-off-by: Chris Chiu <chris.chiu@canonical.com>
Acked-by: Chia-Lin Kao (AceLan) <chia-lin.kao@canonical.com>
Acked-by: Kuan-Ying Lee <kuan-ying.lee@canonical.com>
Signed-off-by: Kuan-Ying Lee <kuan-ying.lee@canonical.com>