thunderbolt: Allow service drivers to specify their own properties
BugLink: https://bugs.launchpad.net/bugs/2160699

The XDomain properties can be useful for service drivers, for example to
implement a registry for the services they expose. So far there has been
no need for service drivers to specify these but with the USB4STREAM
driver that we are going to use them.

This adds remote and local side properties that the service drivers have
access to. Remote side is read-only but the local side can be changed by
a service driver. Also provide a mechanism to notify the remote side
that there are changes.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit abc27e5bfed2fa281ccb23419ef4d1c9ded86398)
Signed-off-by: Chia-Lin Kao (AceLan) <acelan.kao@canonical.com>