ASoC: SDCA: Add manual PM runtime gets to IRQ handlers
BugLink: https://bugs.launchpad.net/bugs/2131725

SDCA interrupt handling is a bit odd, the SDCA IRQ registers are
defined on a device level but the handling of the IRQ is at the
Function level. As such the regmap IRQ's PM runtime operations need to
be on the device itself to ensure those registers are available but an
additional runtime get is required for the Function child when the IRQ
is actually handled. Add the required manual PM runtime gets.

Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.dev>
Link: https://patch.msgid.link/20251020155512.353774-4-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 907364ea3db47530751add6d2d62122ca17329cb)
Signed-off-by: Chris Chiu <chris.chiu@canonical.com>
Acked-by: Chia-Lin Kao (AceLan) <chia-lin.kao@canonical.com>
Signed-off-by: Kuan-Ying Lee <kuan-ying.lee@canonical.com>